requirements.txt
openai==1.85.0
python-dotenv==1.1.0
anthropic>=0.39.0
